Move ExportDataModal
